@@ -1,6 +1,6 @@
 {
   "name": "graylog-web-plugin",
-  "version": "3.3.0-SNAPSHOT",
+  "version": "3.3.0-beta.2-SNAPSHOT",
   "description": "Helper code for streamlining Graylog web interface plugin development",
   "main": "index.js",
   "scripts": {
@@ -52,7 +52,7 @@
     "reflux": "0.2.13",
     "styled-components": "4.4.1",
     "styled-theming": "2.2.0",
-    "webpack": "4.42.1",
+    "webpack": "4.43.0",
     "webpack-cleanup-plugin": "0.5.1",
     "webpack-cli": "3.3.11",
     "webpack-merge": "4.2.2"
