@@ -25,8 +25,8 @@ import org.graylog.plugins.views.migrations.V20190127111728_MigrateWidgetFormatS
 import org.graylog.plugins.views.migrations.V20190304102700_MigrateMessageListStructure;
 import org.graylog.plugins.views.migrations.V20190805115800_RemoveDashboardStateFromViews;
 import org.graylog.plugins.views.migrations.V20191125144500_MigrateDashboardsToViewsSupport.V20191125144500_MigrateDashboardsToViews;
-import org.graylog.plugins.views.migrations.V20191204000000_RemoveLegacyViewsPermissions;
 import org.graylog.plugins.views.migrations.V20191203120602_MigrateSavedSearchesToViewsSupport.V20191203120602_MigrateSavedSearchesToViews;
+import org.graylog.plugins.views.migrations.V20191204000000_RemoveLegacyViewsPermissions;
 import org.graylog.plugins.views.migrations.V20200204122000_MigrateUntypedViewsToDashboards.V20200204122000_MigrateUntypedViewsToDashboards;
 import org.graylog.plugins.views.migrations.V20200409083200_RemoveRootQueriesFromMigratedDashboards;
 import org.graylog.plugins.views.search.SearchRequirements;
@@ -37,6 +37,11 @@ import org.graylog.plugins.views.search.db.SearchJobService;
 import org.graylog.plugins.views.search.db.SearchesCleanUpJob;
 import org.graylog.plugins.views.search.elasticsearch.ESGeneratedQueryContext;
 import org.graylog.plugins.views.search.elasticsearch.ElasticsearchQueryString;
+import org.graylog.plugins.views.search.export.ChunkDecorator;
+import org.graylog.plugins.views.search.export.ExportBackend;
+import org.graylog.plugins.views.search.export.LegacyChunkDecorator;
+import org.graylog.plugins.views.search.export.SimpleMessageChunkCsvWriter;
+import org.graylog.plugins.views.search.export.es.ElasticsearchExportBackend;
 import org.graylog.plugins.views.search.filter.AndFilter;
 import org.graylog.plugins.views.search.filter.OrFilter;
 import org.graylog.plugins.views.search.filter.QueryStringFilter;
@@ -134,6 +139,8 @@ public class ViewsBindings extends ViewsModule {
         registerJacksonSubtype(AutoIntervalDTO.class);
 
         bind(SearchJobService.class).to(InMemorySearchJobService.class).in(Scopes.SINGLETON);
+        bind(ExportBackend.class).to(ElasticsearchExportBackend.class);
+        bind(ChunkDecorator.class).to(LegacyChunkDecorator.class);
 
         registerWidgetConfigSubtypes();
 
@@ -170,6 +177,8 @@ public class ViewsBindings extends ViewsModule {
         queryMetadataDecoratorBinder();
 
         registerExceptionMappers();
+
+        jerseyAdditionalComponentsBinder().addBinding().toInstance(SimpleMessageChunkCsvWriter.class);
     }
 
     private void registerSortConfigSubclasses() {
