@@ -4,15 +4,16 @@ import { useState, useEffect } from 'react';
 import { withRouter } from 'react-router';
 
 import ErrorsActions from 'actions/errors/ErrorsActions';
-import { type ReportedError as ReportedErrorType, ReactErrorType, UnauthorizedErrorType } from 'logic/errors/ReportedErrors';
+import { type ReportedError, ReactErrorType, UnauthorizedErrorType, StreamPermissionErrorType } from 'logic/errors/ReportedErrors';
 
 import RuntimeErrorPage from 'pages/RuntimeErrorPage';
 import UnauthorizedErrorPage from 'pages/UnauthorizedErrorPage';
+import StreamPermissionErrorPage from 'pages/StreamPermissionErrorPage';
 
-const ReportedError = ({ children, router }) => {
-  const [reportedError, setReportedError] = useState<?ReportedErrorType>();
+const ReportedErrorBoundary = ({ children, router }) => {
+  const [reportedError, setReportedError] = useState<?ReportedError>();
 
-  const report = (newError: ReportedErrorType) => setReportedError(newError);
+  const report = (newError: ReportedError) => setReportedError(newError);
 
   useEffect(() => {
     const unlistenErrorsReport = ErrorsActions.report.listen(report);
@@ -36,8 +37,12 @@ const ReportedError = ({ children, router }) => {
     return <UnauthorizedErrorPage error={reportedError.error} />;
   }
 
+  if (reportedError && reportedError.type === StreamPermissionErrorType) {
+    return <StreamPermissionErrorPage error={reportedError.error} />;
+  }
+
   return children;
 };
 
 
-export default withRouter(ReportedError);
+export default withRouter(ReportedErrorBoundary);
