@@ -4,14 +4,14 @@ import styled from 'styled-components';
 
 import { Metric } from 'components/metrics';
 
-const MetricListWrap = styled.ul`
+const MetricListWrap = styled.ul(({ theme }) => `
   padding: 0;
 
   li {
     margin-bottom: 5px;
 
     .prefix {
-      color: #aaa;
+      color: ${theme.color.gray[70]};
     }
 
     .name {
@@ -38,7 +38,7 @@ const MetricListWrap = styled.ul`
     margin-top: 0;
     margin-bottom: 0;
   }
-`;
+`);
 
 class MetricsList extends React.Component {
   static propTypes = {
@@ -48,15 +48,18 @@ class MetricsList extends React.Component {
   };
 
   _formatMetric = (metric) => {
+    const { namespace, nodeId } = this.props;
+
     return (
       <li key={`li-${metric.full_name}`}>
-        <Metric key={metric.full_name} metric={metric} namespace={this.props.namespace} nodeId={this.props.nodeId} />
+        <Metric key={metric.full_name} metric={metric} namespace={namespace} nodeId={nodeId} />
       </li>
     );
   };
 
   render() {
-    const metrics = this.props.names
+    const { names } = this.props;
+    const metrics = names
       .sort((m1, m2) => m1.full_name.localeCompare(m2.full_name))
       .map((metric) => this._formatMetric(metric));
 
