@@ -5,7 +5,7 @@ import ErrorsActions from 'actions/errors/ErrorsActions';
 import StoreProvider from 'injection/StoreProvider';
 import ActionsProvider from 'injection/ActionsProvider';
 // eslint-disable-next-line import/no-cycle
-import { createUnauthorizedError } from 'logic/errors/ReportedErrors';
+import { createFromFetchError } from 'logic/errors/ReportedErrors';
 
 import Routes from 'routing/Routes';
 import history from 'util/History';
@@ -13,7 +13,7 @@ import history from 'util/History';
 export class FetchError extends Error {
   constructor(message, additional) {
     super(message);
-    this.message = message || (additional.message || 'Undefined error.');
+    this.message = message ?? additional?.message ?? 'Undefined error.';
     /* eslint-disable no-console */
     try {
       this.responseMessage = additional.body ? additional.body.message : undefined;
@@ -26,9 +26,37 @@ export class FetchError extends Error {
     /* eslint-enable no-console */
 
     this.additional = additional;
-    this.status = additional.status; // Shortcut, as this is often used
+    this.status = additional?.status; // Shortcut, as this is often used
   }
 }
+
+const reportServerSuccess = () => {
+  const ServerAvailabilityActions = ActionsProvider.getActions('ServerAvailability');
+  ServerAvailabilityActions.reportSuccess();
+};
+
+const defaultOnUnauthorizedError = (error) => ErrorsActions.report(createFromFetchError(error));
+
+const onServerError = (error, onUnauthorized = defaultOnUnauthorizedError) => {
+  const SessionStore = StoreProvider.getStore('Session');
+  const fetchError = new FetchError(error.statusText, error);
+  if (SessionStore.isLoggedIn() && error.status === 401) {
+    const SessionActions = ActionsProvider.getActions('Session');
+    SessionActions.logout(SessionStore.getSessionId());
+  }
+
+  // Redirect to the start page if a user is logged in but not allowed to access a certain HTTP API.
+  if (SessionStore.isLoggedIn() && error.status === 403) {
+    onUnauthorized(fetchError);
+  }
+
+  if (error.originalError && !error.originalError.status) {
+    const ServerAvailabilityActions = ActionsProvider.getActions('ServerAvailability');
+    ServerAvailabilityActions.reportError(fetchError);
+  }
+
+  throw fetchError;
+};
 
 export class Builder {
   constructor(method, url) {
@@ -63,67 +91,43 @@ export class Builder {
       .accept('json')
       .then((resp) => {
         if (resp.ok) {
-          const ServerAvailabilityActions = ActionsProvider.getActions('ServerAvailability');
-          ServerAvailabilityActions.reportSuccess();
+          reportServerSuccess();
           return resp.body;
         }
-
         throw new FetchError(resp.statusText, resp);
-      }, (error) => {
-        const SessionStore = StoreProvider.getStore('Session');
-        if (SessionStore.isLoggedIn() && error.status === 401) {
-          const SessionActions = ActionsProvider.getActions('Session');
-          SessionActions.logout(SessionStore.getSessionId());
+      }, (error) => onServerError(error));
+    return this;
+  }
+
+  file(body, mimeType) {
+    this.request = this.request
+      .send(body)
+      .type('json')
+      .accept(mimeType)
+      .then((resp) => {
+        if (resp.ok) {
+          reportServerSuccess();
+          return resp.text;
         }
-
-        // Redirect to the start page if a user is logged in but not allowed to access a certain HTTP API.
-        if (SessionStore.isLoggedIn() && error.status === 403) {
-          ErrorsActions.report(createUnauthorizedError(error));
-        }
-
-        if (error.originalError && !error.originalError.status) {
-          const ServerAvailabilityActions = ActionsProvider.getActions('ServerAvailability');
-          ServerAvailabilityActions.reportError(error);
-        }
-
-        throw new FetchError(error.statusText, error);
-      });
-
+        throw new FetchError(resp.statusText, resp);
+      }, (error) => onServerError(error));
     return this;
   }
 
   plaintext(body) {
+    const onUnauthorized = () => history.replace(Routes.STARTPAGE);
     this.request = this.request
       .send(body)
       .type('text/plain')
       .accept('json')
       .then((resp) => {
         if (resp.ok) {
-          const ServerAvailabilityActions = ActionsProvider.getActions('ServerAvailability');
-          ServerAvailabilityActions.reportSuccess();
+          reportServerSuccess();
           return resp.body;
         }
 
         throw new FetchError(resp.statusText, resp);
-      }, (error) => {
-        const SessionStore = StoreProvider.getStore('Session');
-        if (SessionStore.isLoggedIn() && error.status === 401) {
-          const SessionActions = ActionsProvider.getActions('Session');
-          SessionActions.logout(SessionStore.getSessionId());
-        }
-
-        // Redirect to the start page if a user is logged in but not allowed to access a certain HTTP API.
-        if (SessionStore.isLoggedIn() && error.status === 403) {
-          history.replace(Routes.STARTPAGE);
-        }
-
-        if (error.originalError && !error.originalError.status) {
-          const ServerAvailabilityActions = ActionsProvider.getActions('ServerAvailability');
-          ServerAvailabilityActions.reportError(error);
-        }
-
-        throw new FetchError(error.statusText, error);
-      });
+      }, (error) => onServerError(error, onUnauthorized));
 
     return this;
   }
@@ -177,6 +181,15 @@ export function fetchPeriodically(method, url, body) {
     .authenticated()
     .noSessionExtension()
     .json(body)
+    .build();
+
+  return queuePromiseIfNotLoggedin(promise)();
+}
+
+export function fetchFile(method, url, body, mimeType = 'text/csv') {
+  const promise = () => new Builder(method, url)
+    .authenticated()
+    .file(body, mimeType)
     .build();
 
   return queuePromiseIfNotLoggedin(promise)();
