@@ -10,7 +10,7 @@ import connect from 'stores/connect';
 import StoreProvider from 'injection/StoreProvider';
 import { ScratchpadProvider } from 'providers/ScratchpadProvider';
 
-import ReportedError from 'components/errors/ReportedError';
+import ReportedErrorBoundary from 'components/errors/ReportedErrorBoundary';
 import RuntimeErrorBoundary from 'components/errors/RuntimeErrorBoundary';
 
 import 'stylesheets/typeahead.less';
@@ -45,7 +45,7 @@ const App = ({ children, currentUser, location }) => {
                   fullName={currentUser.full_name}
                   loginName={currentUser.username}
                   permissions={currentUser.permissions} />
-      <ReportedError>
+      <ReportedErrorBoundary>
         <RuntimeErrorBoundary>
           <ScrollToHint id="scroll-to-hint">
             <Icon name="arrow-up" />
@@ -53,7 +53,7 @@ const App = ({ children, currentUser, location }) => {
           <Scratchpad />
           {children}
         </RuntimeErrorBoundary>
-      </ReportedError>
+      </ReportedErrorBoundary>
     </ScratchpadProvider>
   );
 };
