@@ -1,8 +1,10 @@
 // @flow strict
 import React from 'react';
 import PropTypes from 'prop-types';
+import { withTheme } from 'styled-components';
 import { browserHistory } from 'react-router';
 
+import { type ThemeInterface } from 'theme';
 import Routes from 'routing/Routes';
 import StoreProvider from 'injection/StoreProvider';
 import { isPermitted } from 'util/PermissionsMixin';
@@ -17,7 +19,7 @@ import type { ViewStoreState } from 'views/stores/ViewStore';
 import connect from 'stores/connect';
 import ViewLoaderContext from 'views/logic/ViewLoaderContext';
 import NewViewLoaderContext from 'views/logic/NewViewLoaderContext';
-import CSVExportModal from 'views/components/searchbar/CSVExportModal';
+import CSVExportModal from 'views/components/searchbar/csvexport/CSVExportModal';
 import ShareViewModal from 'views/components/views/ShareViewModal';
 import * as Permissions from 'views/Permissions';
 
@@ -32,6 +34,7 @@ type Props = {
     username: string,
     permissions: Array<string>,
   },
+  theme: ThemeInterface,
 };
 
 type State = {
@@ -48,14 +51,19 @@ const _isAllowedToEdit = (view: View, currentUser = {}) => (
 );
 
 class SavedSearchControls extends React.Component<Props, State> {
-  formTarget: any;
-
   static propTypes = {
     viewStoreState: PropTypes.object.isRequired,
     currentUser: PropTypes.shape({
       username: PropTypes.string.isRequired,
     }).isRequired,
+    theme: PropTypes.shape({
+      color: PropTypes.object,
+    }).isRequired,
   };
+
+  static contextType = ViewLoaderContext;
+
+  formTarget: any;
 
   constructor(props: Props) {
     super(props);
@@ -180,15 +188,12 @@ class SavedSearchControls extends React.Component<Props, State> {
     });
   };
 
-  static contextType = ViewLoaderContext;
-
   render() {
     const { showForm, showList, newTitle, showCSVExport, showShareSearch } = this.state;
-    const { currentUser, viewStoreState } = this.props;
+    const { currentUser, theme, viewStoreState } = this.props;
     const { view, dirty } = viewStoreState;
-
     const csvExport = showCSVExport && (
-      <CSVExportModal closeModal={this.toggleCSVExport} />
+      <CSVExportModal view={view} closeModal={this.toggleCSVExport} />
     );
 
     const shareSearch = showShareSearch && (
@@ -205,7 +210,7 @@ class SavedSearchControls extends React.Component<Props, State> {
     const savedSearchStyle = loaded ? 'star' : 'star-o';
     let savedSearchColor: string = '';
     if (loaded) {
-      savedSearchColor = dirty ? '#ffc107' : '#007bff';
+      savedSearchColor = dirty ? theme.color.variant.warning : theme.color.variant.info;
     }
 
     const disableReset = !(dirty || loaded);
@@ -265,7 +270,7 @@ class SavedSearchControls extends React.Component<Props, State> {
 }
 
 export default connect(
-  SavedSearchControls,
+  withTheme(SavedSearchControls),
   {
     currentUser: CurrentUserStore,
     viewStoreState: ViewStore,
