@@ -1,5 +1,6 @@
 // @flow strict
 import styled, { css, type StyledComponent } from 'styled-components';
+import chroma from 'chroma-js';
 import { util, type ThemeInterface } from 'theme';
 import { Title as NavItemTitle } from './NavItem.styles';
 
@@ -8,7 +9,7 @@ const sidebarWidth = {
   closed: '100%', // width is defined in parent container
 };
 
-export const Container: StyledComponent<{ open: boolean }, ThemeInterface, HTMLDivElement> = styled.div(({ theme }) => css`
+export const Container: StyledComponent<{ open: boolean }, ThemeInterface, HTMLDivElement> = styled.div(({ open, theme }) => css`
   grid-row: 1;
   -ms-grid-row: 1;
   grid-column: 1;
@@ -17,48 +18,47 @@ export const Container: StyledComponent<{ open: boolean }, ThemeInterface, HTMLD
   color: ${util.contrastingColor(theme.color.gray[10], 'AA')};
   height: calc(100vh - 50px);
   padding-top: 20px;
-  position: ${(props) => (props.open ? 'fixed' : 'static')};
+  position: ${open ? 'fixed' : 'static'};
   top: 50px;
-  width: ${(props) => (props.open ? sidebarWidth.open : sidebarWidth.closed)};
+  width: ${open ? sidebarWidth.open : sidebarWidth.closed};
   box-shadow: 3px 0 3px rgba(0, 0, 0, 0.25);
 
   /* z-index is needed for ie11 */
-  z-index: ${(props) => (props.open ? 20 : 'auto')};
+  z-index: ${open ? 20 : 'auto'};
 `);
 
-export const ContentOverlay: StyledComponent<{}, {}, HTMLDivElement> = styled.div`
+export const ContentOverlay: StyledComponent<{}, ThemeInterface, HTMLDivElement> = styled.div(({ theme }) => `
   position: fixed;
   top: 0;
   bottom: 0;
   left: ${sidebarWidth.open};
   right: 0;
-  background: rgba(3, 3, 3, 0.25);
-`;
+  background: ${chroma(theme.color.brand.tertiary).alpha(0.25).css()};
+`);
 
-export const SidebarHeader: StyledComponent<{open: boolean, hasTitle: boolean}, {}, typeof NavItemTitle> = styled(NavItemTitle)`
-  ${(({ open, hasTitle }) => {
-    let justifyContent = 'center';
-    if (open && hasTitle) justifyContent = 'space-between';
-    if (open && !hasTitle) justifyContent = 'flex-end';
-    return `justify-content: ${justifyContent}`;
-  })}
-`;
+export const SidebarHeader: StyledComponent<{open: boolean, hasTitle: boolean}, {}, typeof NavItemTitle> = styled(NavItemTitle)(({ hasTitle, open }) => {
+  const justifyContent = (open && !hasTitle) ? 'flex-end' : 'center';
 
-export const Headline: StyledComponent<{}, {}, HTMLHeadingElement> = styled.h3`
+  return `
+    justify-content: ${(open && hasTitle) ? 'space-between' : justifyContent}
+  `;
+});
+
+export const Headline: StyledComponent<{}, void, HTMLHeadingElement> = styled.h3`
   color: inherit;
   overflow: hidden;
   text-overflow: ellipsis;
   white-space: nowrap;
 `;
 
-export const ToggleIcon: StyledComponent<{}, {}, HTMLDivElement> = styled.div`
+export const ToggleIcon: StyledComponent<{}, void, HTMLDivElement> = styled.div`
   width: 25px;
   text-align: center;
   font-size: 20px;
   cursor: pointer;
 `;
 
-export const HorizontalRuler: StyledComponent<{}, {}, HTMLDivElement> = styled.div`
+export const HorizontalRuler: StyledComponent<{}, void, HTMLDivElement> = styled.div`
   width: 100%;
   padding: 0 10px;
   margin: 5px 0 10px 0;
