@@ -4,6 +4,7 @@ import React from 'react';
 import { CircleMarker, Map, Popup, TileLayer } from 'react-leaflet';
 import chroma from 'chroma-js';
 import { flatten } from 'lodash';
+import createEvent from 'util/CreateEvent';
 
 import style from './MapVisualization.css';
 
@@ -72,7 +73,7 @@ class MapVisualization extends React.Component {
   // Workaround to avoid wrong placed markers or empty tiles if the map container size changed.
   _forceMapUpdate = () => {
     if (this._map) {
-      window.dispatchEvent(new Event('resize'));
+      window.dispatchEvent(createEvent('resize'));
       const { interactive } = this.props;
       this._map.leafletElement.invalidateSize(interactive);
     }
