@@ -1,4 +1,4 @@
 // @flow strict
 import * as Immutable from 'immutable';
 
-export type WidgetMapping = Immutable.Map<string, Immutable.Set<String>>;
+export type WidgetMapping = Immutable.Map<string, Immutable.Set<string>>;
